+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.6.9] - 2025. 05. 21
 
 * Fix text input bug where the first character is replaced when typing the second character ( by @doonfrs )
+* Add enableAutoSelectFirstRow option to automatically select the first row when in selection mode ( by @doonfrs )
 
 ## [1.6.8] - 2025. 04. 15